Time: 0 ms (100.00%) | Memory: 45.3 MB (25.27%) - LeetSync

+++ 2614-maximum-count-of-positive-integer-and-negative-integer/maximum-count-of-positive-integer-and-negative-integer.java
@@ -11,7 +11,7 @@ class Solution {
                 continue;
             }
         }
-        return Math.max(pos,neg);
+        return (int)Math.max(pos,neg);
         
     }
 }
